+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+- Introduce `LockfilePreserver::RubyVersion` [#2][pr2]
+- Introduce `LockfilePreserver::Pipeline` [#2][pr2]
+- Refactor spec fixture originations [#2][pr2]
+- `.keep` now accepts `:ruby_version` to preserve Ruby Version section from the lockfile [#2][pr2]
+- new `.keep_all` method to preserve BUNDLED WITH and RUBY VERSION section from the lockfile [#2][pr2]
+
+[pr2]: https://github.com/jollygoodcode/lockfile_preserver/pull/2
+
 ## 1.0.0 - 2016.03.08
 
 Init LockfilePreserver.